Fixed stack trace concurrency issue
When pausing the debugging process, vscode sometimes didn't select the top call stack entry, because the first stack trace request returned no frames. This happened because it took a while for pushDebugFrames to acquire the lock on a shared thread (which was necessary to unlock after waiting for an unrelated CompletableFuture). The debugger server now uses the java Semaphore, which can be unlocked from a different thread than where it was acquired, allowing pushDebugFrames to acquire the lock without a delay and unlock it after waiting for the CompletableFuture.